🚚 engine.h: Extract `CelSprite` and file loading
