+++ Source/diablo.cpp
@@ -25,6 +25,9 @@
 #include "drlg_l4.h"
 #include "dx.h"
 #include "encrypt.h"
+#include "engine/cel_sprite.hpp"
+#include "engine/load_cel.hpp"
+#include "engine/load_file.hpp"
 #include "error.h"
 #include "gamemenu.h"
 #include "gmenu.h"
