+++ Source/engine/render/cel_render.hpp
@@ -8,6 +8,7 @@
 #include <utility>
 
 #include "engine.h"
+#include "engine/cel_sprite.hpp"
 #include "engine/point.hpp"
 
 namespace devilution {
