+++ Source/engine/render/text_render.cpp
@@ -8,6 +8,7 @@
 #include "DiabloUI/ui_item.h"
 #include "cel_render.hpp"
 #include "engine.h"
+#include "engine/load_cel.hpp"
 #include "engine/point.hpp"
 #include "palette.h"
