+++ Source/engine/render/text_render.hpp
@@ -11,6 +11,7 @@
 
 #include "DiabloUI/ui_item.h"
 #include "engine.h"
+#include "engine/cel_sprite.hpp"
 #include "utils/stdcompat/optional.hpp"
 
 namespace devilution {
