+++ Source/interfac.cpp
@@ -10,6 +10,8 @@
 #include "control.h"
 #include "dx.h"
 #include "engine.h"
+#include "engine/cel_sprite.hpp"
+#include "engine/load_cel.hpp"
 #include "engine/render/cel_render.hpp"
 #include "hwcursor.hpp"
 #include "init.h"
