+++ Source/minitext.cpp
@@ -9,6 +9,8 @@
 #include "control.h"
 #include "dx.h"
 #include "engine.h"
+#include "engine/cel_sprite.hpp"
+#include "engine/load_cel.hpp"
 #include "engine/render/cel_render.hpp"
 #include "engine/render/text_render.hpp"
 #include "textdat.h"
