+++ Source/monster.cpp
@@ -16,6 +16,7 @@
 #include "dead.h"
 #include "drlg_l1.h"
 #include "drlg_l4.h"
+#include "engine/load_file.hpp"
 #include "engine/render/cl2_render.hpp"
 #include "init.h"
 #include "lighting.h"
