+++ Source/player.cpp
@@ -9,6 +9,7 @@
 #include "control.h"
 #include "cursor.h"
 #include "dead.h"
+#include "engine/load_file.hpp"
 #include "gamemenu.h"
 #include "init.h"
 #include "lighting.h"
